Crate cliquenet

Re-exports

pub use error::NetworkError;

Modules

addr

config_builder 🔒

Tools for manipulating the type state of ConfigBuilder.

connection 🔒

delay 🔒

error

metrics 🔒

msg 🔒

net 🔒

noise

queue 🔒

time 🔒

util 🔒

x25519

Structs

Config

ConfigBuilder

Use builder syntax to set the inputs and finish with internal_build().

Network

NetworkReceiver

NetworkSender

SendCommand

Send to peer(s).

SendCommandBuilder

Use builder syntax to set the inputs and finish with build().

Slot

Slots contain messages.

Version

Enums

NetAddr

A network address.

RetryPolicy

Specify if a message should be retried if no ACK is received.

Role

Network peer role.

SendAction

Traits

Metrics

Type that records metrics.