cliquenet::config_builder

Trait State

pub trait State: Sized {
    type Name;
    type Keypair;
    type Bind;
    type Parties;
    type PeerBudget;
    type MaxMessageSize;
    type RetryDelays;
    type MaxRetryDelay;
    type ConnectTimeout;
    type HandshakeTimeout;
    type ReceiveTimeout;
}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type Name

Type state of the member name.

It can implement either IsSet or IsUnset

type Keypair

Type state of the member keypair.

It can implement either IsSet or IsUnset

type Bind

Type state of the member bind.

It can implement either IsSet or IsUnset

type Parties

Type state of the member parties.

It can implement either IsSet or IsUnset

type PeerBudget

Type state of the member peer_budget.

It can implement either IsSet or IsUnset

type MaxMessageSize

Type state of the member max_message_size.

It can implement either IsSet or IsUnset

type RetryDelays

Type state of the member retry_delays.

It can implement either IsSet or IsUnset

type MaxRetryDelay

Type state of the member max_retry_delay.

It can implement either IsSet or IsUnset

type ConnectTimeout

Type state of the member connect_timeout.

It can implement either IsSet or IsUnset

type HandshakeTimeout

Type state of the member handshake_timeout.

It can implement either IsSet or IsUnset

type ReceiveTimeout

Type state of the member receive_timeout.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.