cliquenet::net::peer::peer_builder

Trait State

pub trait State: Sized {
    type Config;
    type Budget;
    type Messages;
    type Inbound;
    type NextSlot;
    type Connection;
}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type Config

Type state of the member config.

It can implement either IsSet or IsUnset

type Budget

Type state of the member budget.

It can implement either IsSet or IsUnset

type Messages

Type state of the member messages.

It can implement either IsSet or IsUnset

type Inbound

Type state of the member inbound.

It can implement either IsSet or IsUnset

type NextSlot

Type state of the member next_slot.

It can implement either IsSet or IsUnset

type Connection

Type state of the member connection.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.