cliquenet

Module net_conf_builder

Tools for manipulating the type state of NetConfBuilder.

See the detailed guide that describes how all the pieces here fit together.

Modules

sealed 🔒

Structs

Empty

Represents a State that has IsUnset implemented for all members.

SetBind

Represents a State that has IsSet implemented for State::Bind.

SetKeypair

Represents a State that has IsSet implemented for State::Keypair.

SetLabel

Represents a State that has IsSet implemented for State::Label.

SetMaxMessageSize

Represents a State that has IsSet implemented for State::MaxMessageSize.

SetMetrics

Represents a State that has IsSet implemented for State::Metrics.

SetName

Represents a State that has IsSet implemented for State::Name.

SetParties

Represents a State that has IsSet implemented for State::Parties.

SetPeerCapacityEgress

Represents a State that has IsSet implemented for State::PeerCapacityEgress.

SetPeerCapacityIngress

Represents a State that has IsSet implemented for State::PeerCapacityIngress.

SetRetryDelays

Represents a State that has IsSet implemented for State::RetryDelays.

SetTotalCapacityEgress

Represents a State that has IsSet implemented for State::TotalCapacityEgress.

SetTotalCapacityIngress

Represents a State that has IsSet implemented for State::TotalCapacityIngress.

Traits

IsComplete

Marker trait that indicates that all required members are set.

IsSet

Marker trait that indicates that the member is set, i.e. at least one of its setters was called.

IsUnset

Marker trait that indicates that the member is unset, i.e. none of its setters was called.

State

Builder’s type state specifies if members are set or not (unset).