cliquenet::net_conf_builder

Trait State

pub trait State: Sized {
    type Name;
    type Label;
    type Keypair;
    type Bind;
    type Parties;
    type TotalCapacityEgress;
    type TotalCapacityIngress;
    type PeerCapacityEgress;
    type PeerCapacityIngress;
    type MaxMessageSize;
    type RetryDelays;
    type Metrics;
}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type Name

Type state of the member name.

It can implement either IsSet or IsUnset

type Label

Type state of the member label.

It can implement either IsSet or IsUnset

type Keypair

Type state of the member keypair.

It can implement either IsSet or IsUnset

type Bind

Type state of the member bind.

It can implement either IsSet or IsUnset

type Parties

Type state of the member parties.

It can implement either IsSet or IsUnset

type TotalCapacityEgress

Type state of the member total_capacity_egress.

It can implement either IsSet or IsUnset

type TotalCapacityIngress

Type state of the member total_capacity_ingress.

It can implement either IsSet or IsUnset

type PeerCapacityEgress

Type state of the member peer_capacity_egress.

It can implement either IsSet or IsUnset

type PeerCapacityIngress

Type state of the member peer_capacity_ingress.

It can implement either IsSet or IsUnset

type MaxMessageSize

Type state of the member max_message_size.

It can implement either IsSet or IsUnset

type RetryDelays

Type state of the member retry_delays.

It can implement either IsSet or IsUnset

type Metrics

Type state of the member metrics.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.