cliquenet

Module retry

Structs

Bucket

Buckets conceptionally contain messages.

Buffer ğŸ”’

Messages are associated with IDs and put into buckets.

Inner ğŸ”’

Message ğŸ”’

Pending ğŸ”’

Data we have received but could not acknowledge yet.

Retry

Retry wraps a Network and returns acknowledgements to senders.

Trailer ğŸ”’

Meta information appended at the end of a message.

Enums

Target ğŸ”’

Constants

MAX_BUCKET

Max. bucket number.

Functions

retry ğŸ”’

Type Aliases

Result ğŸ”’