cliquenet

Struct ConfigBuilder

pub struct ConfigBuilder<S: State = Empty> { /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<S: State> ConfigBuilder<S>

pub fn build(self) -> Config
where S: IsComplete,

Finish building and return the requested object

pub fn name(self, value: &'static str) -> ConfigBuilder<SetName<S>>
where S::Name: IsUnset,

Required.

Network name.

pub fn keypair(self, value: Keypair) -> ConfigBuilder<SetKeypair<S>>
where S::Keypair: IsUnset,

Required.

DH keypair

pub fn bind(self, value: NetAddr) -> ConfigBuilder<SetBind<S>>
where S::Bind: IsUnset,

Required.

Address to bind to.

pub fn parties( self, iter: impl IntoIterator<Item = (PublicKey, NetAddr)>, ) -> ConfigBuilder<SetParties<S>>
where S::Parties: IsUnset,

Required.

Network members with public key and network address.

pub fn peer_budget(self, value: NonZeroUsize) -> ConfigBuilder<SetPeerBudget<S>>
where S::PeerBudget: IsUnset,

Optional (Some / Option setters). Default: NonZeroUsize::new(100).expect("100 > 0").

pub fn maybe_peer_budget( self, value: Option<NonZeroUsize>, ) -> ConfigBuilder<SetPeerBudget<S>>
where S::PeerBudget: IsUnset,

Optional (Some / Option setters). Default: NonZeroUsize::new(100).expect("100 > 0").

pub fn max_message_size( self, value: NonZeroUsize, ) -> ConfigBuilder<SetMaxMessageSize<S>>
where S::MaxMessageSize: IsUnset,

Optional (Some / Option setters). Default: NonZeroUsize::new(10485760).expect("10485760 > 0").

Max. number of bytes per message to send or receive.

pub fn maybe_max_message_size( self, value: Option<NonZeroUsize>, ) -> ConfigBuilder<SetMaxMessageSize<S>>
where S::MaxMessageSize: IsUnset,

Optional (Some / Option setters). Default: NonZeroUsize::new(10485760).expect("10485760 > 0").

Max. number of bytes per message to send or receive.

pub fn retry_delays(self, value: Vec<u8>) -> ConfigBuilder<SetRetryDelays<S>>
where S::RetryDelays: IsUnset,

Optional (Some / Option setters). Default: vec![1, 3, 5, 15, 30].

Retry delays in seconds.

pub fn maybe_retry_delays( self, value: Option<Vec<u8>>, ) -> ConfigBuilder<SetRetryDelays<S>>
where S::RetryDelays: IsUnset,

Optional (Some / Option setters). Default: vec![1, 3, 5, 15, 30].

Retry delays in seconds.

pub fn max_retry_delay( self, value: Duration, ) -> ConfigBuilder<SetMaxRetryDelay<S>>
where S::MaxRetryDelay: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).

pub fn maybe_max_retry_delay( self, value: Option<Duration>, ) -> ConfigBuilder<SetMaxRetryDelay<S>>
where S::MaxRetryDelay: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).

pub fn connect_timeout( self, value: Duration, ) -> ConfigBuilder<SetConnectTimeout<S>>
where S::ConnectTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).

pub fn maybe_connect_timeout( self, value: Option<Duration>, ) -> ConfigBuilder<SetConnectTimeout<S>>
where S::ConnectTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).

pub fn handshake_timeout( self, value: Duration, ) -> ConfigBuilder<SetHandshakeTimeout<S>>
where S::HandshakeTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(10).

pub fn maybe_handshake_timeout( self, value: Option<Duration>, ) -> ConfigBuilder<SetHandshakeTimeout<S>>
where S::HandshakeTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(10).

pub fn receive_timeout( self, value: Duration, ) -> ConfigBuilder<SetReceiveTimeout<S>>
where S::ReceiveTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).

pub fn maybe_receive_timeout( self, value: Option<Duration>, ) -> ConfigBuilder<SetReceiveTimeout<S>>
where S::ReceiveTimeout: IsUnset,

Optional (Some / Option setters). Default: Duration::from_secs(30).