Crate cliquenet

Modules

addr 🔒

config_builder 🔒

Tools for manipulating the type state of ConfigBuilder.

connection 🔒

error

msg 🔒

net 🔒

queue 🔒

time 🔒

x25519

Structs

Config

ConfigBuilder

Use builder syntax to set the inputs and finish with build().

Network

NetworkController

NetworkReceiver

SendCommand

Send to peer(s).

SendCommandBuilder

Use builder syntax to set the inputs and finish with build().

Slot

Slots contain messages.

Enums

NetAddr

A network address.

RetryPolicy

Specify if a message should be retried if no ACK is received.

Role

Network peer role.

SendAction