Crate cliquenet

Re-exports

pub use retry::Retry;

Modules

addr 🔒

chan 🔒

A channel implementation that keeps only a single copy of an item, as identified by its Id.

error 🔒

frame 🔒

Frame header

id 🔒

metrics 🔒

net 🔒

Networking

net_conf_builder 🔒

Tools for manipulating the type state of NetConfBuilder.

retry

time 🔒

x25519 🔒

Structs

Id

InvalidAddress

InvalidKeypair

InvalidPublicKey

InvalidSecretKey

Keypair

NetConf

NetConfBuilder

Use builder syntax to set the inputs and finish with build().

Network

Network is the API facade of this crate.

PublicKey

SecretKey

Enums

Address

A network address.

NetworkError

The various errors that can occur during networking.

Role

Network peer role.

Constants

LAST_DELAY 🔒

MAX_MESSAGE_SIZE

Max. number of bytes for a message (potentially consisting of several frames).

NUM_DELAYS 🔒