cliquenet

Module net

Modules

peer

send_command_builder 🔒

Tools for manipulating the type state of SendCommandBuilder.

server

Structs

Network

NetworkController

NetworkReceiver

SendCommand

Send to peer(s).

SendCommandBuilder

Use builder syntax to set the inputs and finish with build().

Enums

Command 🔒

Server task instructions.

PeerCommand 🔒

Update network peers.

RetryPolicy

Specify if a message should be retried if no ACK is received.

SendAction

Functions

msg_bytes 🔒

Type Aliases

PeerMessage 🔒